+ Toggleable Google Search
Added a new toggle box to search bilibili through google instead of Bilibili directly.
Added a toggle to remove the bilibili site: criteria to search the whole internet for the same CN keywords.
Not Added - Date Selectors to handle google's "before:" and "after:" tags, which could be good for limiting
searches to new OR old runs.
Not Added - more equipment searches like for LWH or Ana etc. Seemed like a whole extra section.